MessageProtocol now uses handlers
