+++ src/lifeblood/worker_pool_message_processor.py
@@ -1,9 +1,10 @@
 from contextlib import contextmanager
-from . import logging
 from .enums import WorkerState
-from .net_messages.tcp_impl.tcp_simple_command_message_processor import TcpCommandMessageProcessor, CommandJsonMessageClient
+from .net_messages.tcp_impl.tcp_simple_command_message_processor import TcpCommandMessageProcessor
+from .net_messages.tcp_impl.clients import CommandJsonMessageClient
 from .net_messages.address import AddressChain
 from .net_messages.messages import Message
+from .net_messages.tcp_impl.message_haldlers import CommandMessageHandlerBase
 
 
 from typing import Optional, Tuple, TYPE_CHECKING
@@ -11,14 +12,10 @@ if TYPE_CHECKING:
     from .simple_worker_pool import WorkerPool
 
 
-class WorkerPoolMessageProcessor(TcpCommandMessageProcessor):
-    def __init__(self, worker_pool: "WorkerPool", listening_address: Tuple[str, int], *, backlog=4096, connection_pool_cache_time=300):
-        super().__init__(listening_address, backlog=backlog, connection_pool_cache_time=connection_pool_cache_time)
-        self.__logger = logging.get_logger('workerpool.message_processor')
+class WorkerPoolMessageHandler(CommandMessageHandlerBase):
+    def __init__(self, worker_pool: "WorkerPool"):
+        super().__init__()
         self.__worker_pool = worker_pool
-
-    async def should_process(self, orig_message: Message):
-        return (await super().should_process(orig_message)) and not self.__worker_pool.is_pool_closed()
 
     def command_mapping(self):
         return {
@@ -42,6 +39,14 @@ class WorkerPoolMessageProcessor(TcpCommandMessageProcessor):
         await client.send_message_as_json({
             'ok': True
         })
+
+
+class WorkerPoolMessageProcessor(TcpCommandMessageProcessor):
+    def __init__(self, worker_pool: "WorkerPool", listening_address: Tuple[str, int], *, backlog=4096, connection_pool_cache_time=300):
+        super().__init__(listening_address,
+                         backlog=backlog,
+                         connection_pool_cache_time=connection_pool_cache_time,
+                         message_handlers=(WorkerPoolMessageHandler(worker_pool),))
 
 
 #
